+++ i18n/concavehull_de.ts
@@ -1,112 +1,238 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!DOCTYPE TS>
-<TS version="2.0" language="DE">
+<TS version="2.0" language="de">
 <context>
     <name>ConcaveHull</name>
     <message>
-        <location filename="../concavehull.py" line="480"/>
+        <location filename="../concave_hull.py" line="707"/>
         <source>No shapefile name specified</source>
         <translation>Keine Shapedatei angegeben</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="282"/>
+        <location filename="../concave_hull_dialog_base.ui" line="336"/>
         <source>ConcaveHull</source>
         <translation>ConcaveHull</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="32"/>
+        <location filename="../concave_hull_dialog_base.ui" line="32"/>
         <source>Specify the number of nearest neighbors to start with</source>
-        <translation>Mindestgröße der Nachbarschaft</translation>
+        <translation>Mindestgröße der Nachbarschaft zu Beginn der Berechnung</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="35"/>
+        <location filename="../concave_hull_dialog_base.ui" line="35"/>
         <source>Number of neighbors</source>
         <translation>Anzahl der Nachbarpunkte</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="88"/>
+        <location filename="../concave_hull_dialog_base.ui" line="88"/>
         <source>Check and specify a neighborhood size to cluster the data prior constructing a hull</source>
         <translation>Mindestgröße der Cluster angeben</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="91"/>
+        <location filename="../concave_hull_dialog_base.ui" line="91"/>
         <source>Find SNN clusters</source>
         <translation>SNN Cluster bilden</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="136"/>
+        <location filename="../concave_hull_dialog_base.ui" line="236"/>
         <source>Neighborhood size</source>
         <translation>Mindestgröße der Cluster</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="217"/>
+        <location filename="../concave_hull_dialog_base.ui" line="249"/>
         <source>Input layers</source>
         <translation>Eingabelayer</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="226"/>
+        <location filename="../concave_hull_dialog_base.ui" line="258"/>
         <source>Select one or more input layers</source>
         <translation>Selektiere einen oder mehrere Layer</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="236"/>
+        <location filename="../concave_hull_dialog_base.ui" line="280"/>
         <source>Use only selected features on input layers</source>
-        <translation>Nur selektierte Objekte verwenden</translation>
+        <translation>Nur selektierte Objekte der Eingabelayer verwenden</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="239"/>
+        <location filename="../concave_hull_dialog_base.ui" line="283"/>
         <source>Use selected features only</source>
         <translation>Nur selektierte Objekte verwenden</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="249"/>
+        <location filename="../concave_hull_dialog_base.ui" line="303"/>
         <source>Output layer</source>
         <translation>Ausgabelayer</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="255"/>
+        <location filename="../concave_hull_dialog_base.ui" line="309"/>
         <source>New memory layer</source>
         <translation>Neuen Memorylayer anlegen</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="279"/>
+        <location filename="../concave_hull_dialog_base.ui" line="333"/>
         <source>Type in the name of a new memory layer</source>
         <translation>Gebe den Namen des neuen Memorylayer an</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="291"/>
+        <location filename="../concave_hull_dialog_base.ui" line="345"/>
         <source>Existing layer</source>
         <translation>Bestehenden Layer verwenden</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="312"/>
+        <location filename="../concave_hull_dialog_base.ui" line="366"/>
         <source>Choose an existing polygon layer from the list</source>
         <translation>Wähle einen bestehenden Polygonlayer aus der Liste</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="327"/>
+        <location filename="../concave_hull_dialog_base.ui" line="381"/>
         <source>New shapefile</source>
         <translation>Neue Shapedatei erstellen</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="364"/>
+        <location filename="../concave_hull_dialog_base.ui" line="418"/>
         <source>Browse for the location of a new or existing Shape file</source>
-        <translation>Neue Shapedatei angeben oder eine bestehende Datei auswählen</translation>
+        <translation>Eine bestehende Datei im Explorer auswählen</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="367"/>
+        <location filename="../concave_hull_dialog_base.ui" line="421"/>
         <source>Browse ...</source>
-        <translation>Durchsuchen</translation>
+        <translation>Durchsuchen ...</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="376"/>
+        <location filename="../concave_hull_dialog_base.ui" line="430"/>
         <source>Add the new layer to the map</source>
         <translation>Ergebnis der Karte hinzufügen</translation>
     </message>
     <message>
-        <location filename="../ui_concavehull.ui" line="379"/>
+        <location filename="../concave_hull_dialog_base.ui" line="433"/>
         <source>Add result to map</source>
         <translation>Ergebnis der Karte hinzufügen</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull_dialog_base.ui" line="119"/>
+        <source>Output points that are not part of any cluster</source>
+        <translation>Punkte, die keinem Cluster angehören</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull_dialog_base.ui" line="112"/>
+        <source>Output noise points</source>
+        <translation>Streupunkte ausgeben</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull_dialog_base.ui" line="122"/>
+        <source>Output clusters only</source>
+        <translation>Nur Clusterausgeben</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull_dialog_base.ui" line="290"/>
+        <source>If grouping by attribute is selected, the algorithm will group the features in the input layer using unique values in that field and generate individual polygons in the output layer for each group</source>
+        <translation>Wenn ein Attribut für die Gruppierung angegeben wird, werden Cluster für jede gruppe unabhängig voneinander gebildet</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull_dialog_base.ui" line="293"/>
+        <source>Group by attribute</source>
+        <translation>Nach einem Attribut gruppieren</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="691"/>
+        <source>&amp;Concave Hull</source>
+        <translation>&amp;Konkave Hülle</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="682"/>
+        <source>Concave Hull</source>
+        <translation>Konkave Hülle</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="921"/>
+        <source>Error writing shapefile: </source>
+        <translation>Schreibfehler bei Shapefile: </translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1129"/>
+        <source>Empty point set for group {}. Excluded from processing</source>
+        <translation>Keine Eingabepunkte für Gruppe {} gefunden. Überspringe</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1139"/>
+        <source>A minimum of 3 points is required ({} are given for group {}). Excluded from processing</source>
+        <translation>Mindestens 3 Punkte sind erforderlich (nur {} vorhanden für Gruppe {}). Überspringe</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1148"/>
+        <source>{} neighbors are required, {} input points are given for group {}. 
+                        Excluded from processing</source>
+        <translation>{} Nachbarpunkte sind mindestens erforderlich, {} vorhanden für Gruppe {}. Überspringe</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1195"/>
+        <source>Empty point set. Processing aborted</source>
+        <translation>Keine Eingabepunkte vorhanden. Abbruch</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1203"/>
+        <source>A minimum of 3 points is required ({} are given). Processing aborted</source>
+        <translation>Mindestens 3 Punkte sind erforderlich ({} vorhanden). Abbruch</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1212"/>
+        <source>{} neighbors are required, {} input points are given. Processing aborted</source>
+        <translation>{} Nachbarpunkte sind mindestens erforderlich, nur {} vorhanden. Abbruch</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1220"/>
+        <source>Please be patient, processing of more then {} points may take a while</source>
+        <translation>Bitte habe Geduld, die Verarbeitung von {} Punkten kann eine Weile dauern</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1227"/>
+        <source>Please confirm</source>
+        <translation>Bestätige</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1227"/>
+        <source>Do you really want to proceed?</source>
+        <translation>Möchtest du den Prozess fortsetzen?</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1365"/>
+        <source>Error creating concave hull for group {}</source>
+        <translation>Fehler bei Konstruktion der konkaven Hülle für Gruppe {}</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1368"/>
+        <source>Error creating concave hull</source>
+        <translation>Fehler bei Konstruktion der konkaven Hülle</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1326"/>
+        <source>{} noise points not added to any cluster</source>
+        <translation>{} Streupunkte konnten keinem Cluster zugeordnet werden</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1342"/>
+        <source>Clustered points layer created</source>
+        <translation>Ausgabelayer für Punktcluster erzeugt</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1346"/>
+        <source>Creating clustered points layer failed</source>
+        <translation>Erstellen des Ausgabelayers für Cluster fehlgeschlagen</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1379"/>
+        <source>{} Concave hulls created successfully</source>
+        <translation>{} konkave Hülle(n) erfolgreich erzeugt</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1383"/>
+        <source>No concave hull was created</source>
+        <translation>Keine konkave Hülle erzeugt</translation>
+    </message>
+    <message>
+        <location filename="../concave_hull.py" line="1387"/>
+        <source>Creating/editing concave hulls layer failed</source>
+        <translation>Erstellen des Ausgabelayers für konkave Hüllen fehlgeschlagen</translation>
     </message>
 </context>
 </TS>